# Speedier `Sqids.new` 🦑