With the current implementation a call to `Sqid.new` can take ~3/4ms.

This improves this in two ways:

1. If the Default alphabet and blocklist are used, skip filtering, as they already meet the rules

    That gets `Sqid.new` down from ~3ms to 0.07ms

2. When not using defaults, move `alphabet.downcase` out of the `select` loop, so we don't call it once per block list item.

   That gets us from 3.315ms to 0.95ms 🥳

Numbers and benchmarking approach in this issue here:
Fixes #6