@4kimov 2024-07-25
Thank you for the optimization 🙏
Released in [v0.2.1](https://rubygems.org/gems/sqids/versions/0.2.1)